refactor: separate language change functionality into a new ChangeLanguage component and update AppBar for improved clarity
