+++ src/app/[locale]/_components/AppBar.tsx
@@ -1,21 +1,10 @@
-"use client";
-
-import { Button } from "components/atoms";
-import { ButtonGroup } from "components/molecules";
 import { CodeIcon, GithubIcon, LinkedInIcon } from "components/icons";
-import {
-  useChangeLocale,
-  useCurrentLocale,
-  useScopedI18n,
-} from "locales/client";
 import { merge } from "utils/clsx";
+import ChangeLanguage from "./ChangeLanguage";
+import { getScopedI18n } from "locales/server";
 
-const languages = ["en", "es"];
-
-const AppBar = () => {
-  const locale = useCurrentLocale();
-  const t = useScopedI18n("common");
-  const changeLocale = useChangeLocale({ preserveSearchParams: true });
+const AppBar = async () => {
+  const t = await getScopedI18n("common");
 
   return (
     <div
@@ -27,20 +16,14 @@ const AppBar = () => {
     >
       <div className="container m-x-auto flex items-stretch gap-3">
         <a
-          href={`/${locale}`}
+          href={`/|`}
           className="flex items-center"
           aria-label={t("appBar.ariaLabel.home")}
         >
           <CodeIcon />
         </a>
         <span className="grow" />
-        <ButtonGroup size="small" value={locale} onChange={changeLocale}>
-          {languages.map((locale) => (
-            <Button key={locale} value={locale}>
-              {locale}
-            </Button>
-          ))}
-        </ButtonGroup>
+        <ChangeLanguage />
 
         <a
           href="https://www.linkedin.com/in/juanmarcon"
